+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@ For git tags, official relases are marked as "v1.0.0", while unofficial ones a m
 * Commands included in ``Makefile`` where transcripted in ``composer.json``.
 * [Docker](https://www.docker.com/) integration! It is done through ``Dockerfile``, ``docker-compose.yml``, ``.dockerignore`` and ``.env`` files.
 * You can click on a read icon in order to read a word.
+* Text-To-Speech (TTS) Settings!
+  * Change the language rate, pitch, or local region for text-to-speech.
 
 ### Changed
 * mbstring and mysqli extensions are now clearly asked for by Composer.